Move post JDK 8 tests to new DoubleTestOnJDK20.scala
