+++ unit-tests/shared/src/test/scala/org/scalanative/testsuite/javalib/lang/DoubleTest.scala
@@ -4,7 +4,7 @@ import java.lang.Double.{
   doubleToLongBits, doubleToRawLongBits, longBitsToDouble, toHexString
 }
 import java.lang._
-// Because this test is in the java.lang package, an unqualified Double
+// Because this test is the java.lang package, an unqualified Double
 // is a java.lang.Double. Prior art used unqualified Double freely,
 // with that intent. Scala.js JDouble is introduced to minimize changes
 // in ported Scala.js tests. Existing usages of unqualified Double
@@ -423,29 +423,5 @@ class DoubleTest {
       -scala.Double.NegativeInfinity,
       delta
     )
-  }
-
-  /* Scala Native additions for features added after Java 8.
-   */
-
-  /** Since: Java 12 */
-  @Test def testDescribeConstable(): Unit = {
-    val expected = JDouble.valueOf(1.618)
-    val result = expected
-      .describeConstable()
-      .orElseGet(() => { fail("describeConstable is empty"); null })
-
-    assertTrue(result.eq(expected)) // Require reference equality
-  }
-
-  /** Since: Java 12 */
-  /** Requires reflection so can not presently (SN 0.5.12) be implemented on
-   *  Scala Native.
-   */
-  //  @Test def testresolveConstantDesc (): Unit
-
-  /** Since: Java 19 */
-  @Test def testPrecision(): Unit = {
-    assertEquals("PRECISION", 53, JDouble.PRECISION)
   }
 }
